Existing users (by email) can now login with Facebook
Still something to do but all seem working just fine. Existing users
are now able to login using Facebook: their email is fetched and
compared with our database and if present, then their account is linked
with their facebook one, and they are logged in. If they have already
logged in before, they are logged in immediately.