Removed the online presence check as it wasn't working properly. Fixed #1
Before: it checked for the user data and if the system would have been
able to get them, then the Facebook login process wasn't performed to
speed it up. This led into #1.

Now: every login is passed to Facebook which checks itself for online
presence of the user and acts dependently on it.