Fixed protected results being showed to users
This bug showed a lot of protected informations if the user didn't
authorize the app to share his data, including the app Key and Secret.
This would lead into a small lack of security: users would be able to
use the key and the secret for their board, but they would be unuseful
because of Facebook's strong privacy which restrict the use of an app
for a single domain specified by the developer.